+++ lib/allgood.rb
@@ -3,6 +3,7 @@
 require_relative "allgood/version"
 require_relative "allgood/engine"
 require_relative "allgood/configuration"
+require_relative "allgood/cache_store"
 
 module Allgood
   class Error < StandardError; end